Creación del componente input-nombre
