--- src/app/app.component.css
@@ -1,7 +1,0 @@
-.rojo{
-  color: red;
-}
-
-.azul{
-  color:blue;
-}
