+++ src/app/components/saludo-input-nombre/saludo-input-nombre.component.html
@@ -1,0 +1,8 @@
+
+<h3 class = 'titulo'>Cual es tu nombre?</h3>
+<form  class="formulario">
+
+  <input type="text" placeholder="Nombre"  [(ngModel)]="nombre" [ngModelOptions]="{standalone: true}" >
+  <button (click)="emitirNombre()">Enviar</button>
+</form>
+<p style="color: red;">{{mensajeDeError}}</p>
