+++ src/app/components/saludo-texto/saludo-texto.component.html
@@ -1,0 +1,4 @@
+
+<h2 class = 'titulo' *ngIf="mostrarSaludo()" [ngClass]="colorSaludo()">
+  ¡Hola, {{nombre()}}!
+</h2>
