+++ src/app/pages/despedida-page/despedida-page.component.html
@@ -1,7 +1,7 @@
 
 <div class = 'contenedor'>
 
-  <h1>Adiós, {{nombre}} ! </h1>
+  <h2 class = 'titulo '>Adiós, {{nombre}} ! </h2>
 
   <div class = 'botones'>
     <button (click)="saludar()">Volver</button>
