--- src/app/pages/saludo/saludo-texto/saludo-texto.component.html
@@ -1,4 +1,0 @@
-
-<h1 *ngIf="mostrarSaludo()" [ngClass]="colorSaludo()">
-  ¡Hola, {{nombre()}}!
-</h1>
