--- src/app/pages/saludo/saludo.component.css
@@ -1,53 +1,0 @@
-
-.contenedor {
-  max-width: 500px;
-  margin: 50px auto;
-  padding: 20px;
-  text-align: center;
-  border-radius: 12px;
-  box-shadow: 0 0 15px rgba(0, 0, 0, 0.1);
-  background-color: #f9f9f9;
-}
-
-.saludo {
-  font-size: 2rem;
-  margin-bottom: 20px;
-  transition: color 0.3s ease;
-}
-
-.rojo {
-  color: red;
-}
-
-.verde {
-  color: green;
-}
-
-.botones {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  gap: 10px;
-}
-
-button {
-  padding: 10px 15px;
-  border: none;
-  border-radius: 6px;
-  background-color: #ddd;
-  cursor: pointer;
-  transition: background-color 0.2s ease;
-}
-
-button:hover {
-  background-color: #bbb;
-}
-
-
-.imagen-saludo {
-  display: block;
-  margin: 30px auto 0;
-  width: 400px;
-  max-width: 90%;
-  border-radius: 10px;
-}
