--- src/app/pages/saludo/saludo.component.html
@@ -1,31 +1,0 @@
-
-<div class = 'contenedor'>
-
-
-  <app-saludo-texto
-    [mostrarSaludo]="mostrarSaludo"
-    [colorSaludo]="colorSaludo"
-    [nombre]="nombre | mayuscula"
-  ></app-saludo-texto>
-
-  <app-saludo-botones
-    [mostrarSaludo]="mostrarSaludo"
-    (mostrarSaludoChange)="mostrarSaludo = $event"
-
-    [colorSaludo]="colorSaludo"
-    (colorSaludoChange)="colorSaludo=$event"
-
-    [nombre]="nombre | mayuscula"
-    (nombreChange)="nombre = $event"
-  > </app-saludo-botones>
-
-</div>
-
-
-<ng-container *ngIf="mostrarSaludo; else imagenSinSaludo">
-  <img src="/saludo.png" alt="Saludo animado" class="imagen-saludo" />
-</ng-container>
-
-<ng-template #imagenSinSaludo>
-  <img src="/noSaludo.png" alt="Chica confundida" class="imagen-saludo" />
-</ng-template>
